Add a higher level wrapper around precise_time_ns()
Since precise_time_ns() returns a number of nanoseconds since an
unspecified epoch, the set of operations that can be performed is a bit
limited. The epoch could theoretically be far enough in the past that
the value overflows during the runtime of the program, so the values
can't even necessarily be compared via the normal ordering operators.

The `PreciseTimestamp` type is an opaque wrapper around the precise
time, which only provides operations that actually make sense to use.
Right now that's only a method that turns two `PreciseTimestamps` into
the `Duration` of time between them. Since we care about overflow, we
*assume* that the `later` timestamp was created after the `self`
timestamp, and the result is positive no matter what (unless you waited
more than 293 years between measurements).