Add a more robust monotonic time wrapper
It looks like overflow isn't really a thing worth worrying about.
Windows guarantees 100 years of uptime before overflow of
QueryPerformanceCounter, and the OSX and Linux timers appear to be based
off of boot time as well.

Storing the time as the source data type makes the implementation more
complex, but avoids precision loss or shifts due to conversions to the
greatest extent possible.