Implement formatting
With language support, currently English and Spanish. Other languages
are trivial to add.

DeferredFormat will be removed before a release; it was originally a
design decision that no longer makes sense.

In the future, the parsing of the format string could be performed at
compile-time. Currently, this would require some trickery for complete
type-checking, as type-based macros do not exist.