Re-implement Duration using primitives
This simplifies the implementation in some places, due to not having to
constantly convert back and forth to the type in std. As an added bonus,
a number of methods are now const, and the size of the struct decreased
from 24 to 16 bytes.

Though some of the method signatures  change, this is necessary. Under
the old implementation, it was not possible to directly create a value
between 0 and -1 second.