MSRV of 1.34.0, bug fixes, move to alloc feature
Previously, some components were not checked when parsing, even when
they should have been. This was more of a problem because of the use of
unchecked methods. All final parsing methods use the checked versions,
returning a component error if failing.

Error handling from parsing could/probably should be reworked in the
future at some point.

The "std" feature has been renamed to "alloc", which is the negative
equivalent. This allows for use of serde with the standard library
enabled in 1.34. This is necessary because cargo currently doesn't
support disabling a feature with another feature.