Fix behavior for `OffsetDateTime`
The previous behavior was ambiguous, as it was unclear whether the
provided offset was the one being assumed and being converting to UTC or
vice versa.

This ambiguity extended to within the code, so there were some bugs
where the behavior was backwards. These have been fixed, as it's now
clearly documented that the datetime stored in the struct is the UTC
representation, not local.