Enable building of documentation on stable
By default, we will build documentation without any nightly features
enabled. We can then use an opt-in feature (`__doc`) to enable the
nightly feature gate. This flag is used when building for docs.rs and
time-rs.github.io. It should never be relied upon externally, as it can
be removed at any time (and will be after the `doc_cfg` feature is
stabilized).