Method to get local UTC offset at a given datetime
This implementation should work on Windows, Mac, Linux, and Solaris. As
Redox is *nix, it should work there as well.

`OffsetDateTime::now()` and `impl From<SystemTime> for OffsetDateTime`
both return a value with the local offset.

This necessarily requires syscalls, and as such
`#![forbid(unsafe_code)]` has been changed to `#![deny(unsafe_code)]`.
It must be explicitly `#[allow]`ed in every location it is used, along
with a general message describing why the usage is safe.

Co-authored-by: Hroi Sigurdsson <hroi@asdf.dk>