Better handling of cfgs
As cfgs are not unique to a given crate, they must be disambiguated to
avoid conflicts. This is accomplished by avoiding them where possible
and using a crate-version prefixed name otherwise.

docs.rs has an option to use Cargo metadata to pass additional flags.
This can be used for docs.rs, but building documentation for
time-rs.github.io needs the flags passed via an environment variable.
This is done in CI.