Reduce uploaded crate size by 37%
By splitting nearly every test into a separate tests directory, they can
be omitted from the crates.io upload. As a bonus, it helps ensure that
any changes to APIs are wholly internal. This would have caught a couple
errors in the past.

As part of this change, I've also run `cargo diet`, which excludes any
file (in this case including the tests directory) that isn't needed on
crates.io. This eliminates things like rustfmt, the logo, and CI config.