Improve test coverage
Cover errors and serde. Rand coverage was improved with a new seed, and
Date now has its debug implementation checked.

As part of this coverage expansion, it was discovered that only one
implementation of a generic function was being used internally. This
generic type was removed in favor of concrete types, which slightly
reduces the amount of code present.