Revamped macros
The macros have been rewritten from scratch. To avoid dealing with
proc_macro's limited API, all macros now take a string and parse that.

As part of this change, the macros no longer depend on syn, quote, or
proc-macro2. A `datetime!` macro was also introduced, equivalent to a
date followed immediately by a time.