Add weekday value to `iso_year_week` return
The method has been renamed from `iso_year_week` to `to_iso_week_date`
to match this change. The old method is still present for internal
usage.

Note that this and the previous commit still need to be applied to
`PrimitiveDateTime` and `OffsetDateTime`, as well as their respective
tests, documentation, and benchmarks.