Revamp `UtcOffset` API
Store the components directly, rather than as the total number of
seconds from UTC. The API has been massively simplified, such that the
two primary methods are `from_hms` and `as_hms`. `to_seconds` replaces
`as_seconds`, and is intended to be used as a convenience method for
those who need access to that specific piece of information.