Introduce `large-dates` feature
This allows for optimization based on based on the smaller valid range.
While the smaller range of years fits well within an i16, it is still an
i32 in method headers for API consistency.

This results in performance improvements of ~15% for all arithmetic,
with ~40% gains for `Date::from_julian_day`.