Check feature powerset on all supported versions
This will run three times per week. It will ensure that any combination
of public feature flags compiles on any of the supported Rust versions.
The test suite is not run because the process may run out of storage on
the virtual machine.

This is not run on every push due to the duration these checks will
take.