Update dependencies
This upgrades quickcheck to 1.0 and rand to 0.8, in addition to bumping
patch versions crate-wide.

From my reading of quickcheck's documentation, there is no reason to use
`Gen::size` here, as that's intended for values without a statically
known size (namely heap-allocated ones); that is not the case here. In
response, such tests have been removed in their entirety.

Quickcheck now has `#![no_std]` support, though that's probably not
particularly useful on its own.