Handle modifiers, revamp combinators
The combinators are much more targeted, such that writing an
intermediate parser is quite easy. There is also a helper method to
avoid needless repetition where the input is shadowed to defer mutation.

The marker trait `Integer` has been restricted to signed integers to
reflect the fact that the sign is not parsed as part of it. This ensures
that the caller is conscious of the fact that they need to change the
sign themselves.