Update benchmarks
Eliminate benchmarks of any items that are wholly redundant, namely
items that are simply delegated to an internal field or method that is
already benchmarked. Also eliminated are trivial implementations for
`Instant`, as it is guaranteed to be a thin wrapper around the stdlib
equivalent. Many local variables that were `const` (due to the macros
simulating const blocks) have been inlined.