Unify benchmarks, move to near-standard location
This avoids the issue of requiring two additional crates for the sole
purpose of benchmarking. It is now possible to perform benchmarking via
the standard `cargo bench` without having to change directories first.

It is not currently possible to use the standardized "benches" name, as
cargo then looks at each file individually. As we want the benchmarks to
be unified, we provide a single entry point and direct cargo to that.