Introduce internal `quote!` macro
Similar to the macro from the quote crate, but tailored to the built-in
proc_macro crate. Interpolation is different as well, but overall still
yields a great improvement over manually writing it out. The only
situation where this macro is not used is where the `Span` needs to be
set.