Add cache for docs, coverage, remove for powerset
While the docs and coverage will only have their cache valid for a
single nightly, it's better than nothing and will speed up CI when there
are multiple pushes in a day.

For the scheduled tasks, the cache is quite large and provides minimal
benefit relative to the runtime of the process. It is also run
asynchronously, so a person doesn't typically need to wait for its
output.