Make all modifiers `#[non_exhaustive]`
This necessitates some hacks to work around current compiler
limitations. Specifically, const trait impls are not permitted, so a new
`default` method is introduced that is const. Once they are permitted,
this new implementation can be removed without being a breaking change
because `Default` is in the prelude.

This commit also introduces a smaller change wherein
`OffsetHour::sign_is_mandatory` now defaults to `true`. This matches the
behavior of the parser.