Update use of libc::timespec to prepare for future libc version
In a future release of the `libc` crate, `libc::timespec` will contains
private padding fields on 32-bit `*-linux-musl` targets and so the
struct will no longer be able to be created using the literal
initializer syntax.

The only uses in this crate of `libc::timespec` in this way were zero
initializing the struct. Thus, these can be replaced by a call to
`std::mem::zeroed()` which is compatible with both current versions of
the `libc` crate as well as the future version which will contain those
private padding fields.